update: LSPの設定を移動
